Create ProfileAPIController

+++ profile-service/src/main/java/project/profileservice/domain/Profile.java
@@ -20,7 +20,7 @@ public class Profile {
     private Long user_id;
     private int nowStrick;
     private int maxStrick;
-    private Long point;
+    private double point;
 
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "profile")
     private List<ProfileBadge> profileBadges = new ArrayList<ProfileBadge>();
@@ -67,8 +67,8 @@ public class Profile {
     /**
      * point를 업데이트 한다.
      */
-    public void updatePoint(Long point) {
-        Long restPoint = this.point + point;
+    public void updatePoint(double point) {
+        double restPoint = this.point + point;
         if (restPoint < 0) {
             throw new NotEnoughPointException("Cannot update point");
         }
